feedback / reorg tests
