+++ synapse/tests/test_model_infotech.py
@@ -8,6 +8,7 @@ import synapse.models.crypto as s_m_crypto
 
 import synapse.tests.files as s_t_files
 import synapse.tests.utils as s_t_utils
+import synapse.tests.test_lib_scrape as s_t_scrape
 
 class InfotechModelTest(s_t_utils.SynTest):
 
@@ -28,28 +29,6 @@ class InfotechModelTest(s_t_utils.SynTest):
             self.eq(nodes[0].get('desc'), 'omgwtfbbq')
             self.eq(nodes[0].get('url'), 'https://cwe.mitre.org/data/definitions/120.html')
             self.eq(nodes[0].get('parents'), ('CWE-119',))
-
-            self.eq(r'foo:bar', core.model.type('it:sec:cpe').norm(r'cpe:2.3:a:foo\:bar:*:*:*:*:*:*:*:*:*')[1]['subs']['vendor'])
-
-            with self.raises(s_exc.BadTypeValu):
-                nodes = await core.nodes('[it:sec:cpe=asdf]')
-
-            with self.raises(s_exc.BadTypeValu):
-                nodes = await core.nodes('[it:sec:cpe=cpe:2.3:1:2:3:4:5:6:7:8:9:10:11:12]')
-
-            nodes = await core.nodes('[ it:sec:cpe=cpe:2.3:vertex:synapse ]')
-            self.eq(nodes[0].ndef, ('it:sec:cpe', 'cpe:2.3:vertex:synapse:*:*:*:*:*:*:*:*:*'))
-
-            nodes = await core.nodes('''[
-                it:sec:cpe=cpe:2.3:a:microsoft:internet_explorer:8.0.6001:beta:*:*:*:*:*:*
-            ]''')
-            self.len(1, nodes)
-            self.eq(nodes[0].ndef, ('it:sec:cpe', 'cpe:2.3:a:microsoft:internet_explorer:8.0.6001:beta:*:*:*:*:*:*'))
-            self.eq(nodes[0].get('part'), 'a')
-            self.eq(nodes[0].get('vendor'), 'microsoft')
-            self.eq(nodes[0].get('product'), 'internet_explorer')
-            self.eq(nodes[0].get('version'), '8.0.6001')
-            self.eq(nodes[0].get('update'), 'beta')
 
             nodes = await core.nodes('''[
                 it:mitre:attack:group=G0100
@@ -1520,6 +1499,28 @@ class InfotechModelTest(s_t_utils.SynTest):
     async def test_infotech_cpes(self):
 
         async with self.getTestCore() as core:
+            self.eq(r'foo:bar', core.model.type('it:sec:cpe').norm(r'cpe:2.3:a:foo\:bar:*:*:*:*:*:*:*:*:*')[1]['subs']['vendor'])
+
+            with self.raises(s_exc.BadTypeValu):
+                nodes = await core.nodes('[it:sec:cpe=asdf]')
+
+            with self.raises(s_exc.BadTypeValu):
+                nodes = await core.nodes('[it:sec:cpe=cpe:2.3:1:2:3:4:5:6:7:8:9:10:11:12]')
+
+            nodes = await core.nodes('[ it:sec:cpe=cpe:2.3:vertex:synapse ]')
+            self.eq(nodes[0].ndef, ('it:sec:cpe', 'cpe:2.3:vertex:synapse:*:*:*:*:*:*:*:*:*'))
+
+            nodes = await core.nodes('''[
+                it:sec:cpe=cpe:2.3:a:microsoft:internet_explorer:8.0.6001:beta:*:*:*:*:*:*
+            ]''')
+            self.len(1, nodes)
+            self.eq(nodes[0].ndef, ('it:sec:cpe', 'cpe:2.3:a:microsoft:internet_explorer:8.0.6001:beta:*:*:*:*:*:*'))
+            self.eq(nodes[0].get('part'), 'a')
+            self.eq(nodes[0].get('vendor'), 'microsoft')
+            self.eq(nodes[0].get('product'), 'internet_explorer')
+            self.eq(nodes[0].get('version'), '8.0.6001')
+            self.eq(nodes[0].get('update'), 'beta')
+
             cpe23 = core.model.type('it:sec:cpe')
             cpe22 = core.model.type('it:sec:cpe:v2_2')
 
@@ -1591,6 +1592,16 @@ class InfotechModelTest(s_t_utils.SynTest):
 
                 norm, info = cpe22.norm(v2_2)
                 self.eq(norm, _cpe22)
+
+    async def test_cpe_scrape_one_to_one(self):
+
+        async with self.getTestCore() as core:
+            q = '[it:sec:cpe=$valu]'
+            for _, valu in s_t_scrape.s_scrape.scrape(s_t_scrape.cpedata, ptype='it:sec:cpe'):
+                nodes = await core.nodes(q, opts={'vars': {'valu': valu}})
+                self.len(1, nodes)
+                node = nodes[0]
+                self.eq(node.ndef[1], valu.lower())
 
     async def test_infotech_c2config(self):
         async with self.getTestCore() as core:
